% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_plot_for_legend.R
\name{empty_plot_for_legend}
\alias{empty_plot_for_legend}
\title{Reset the plotting window}
\usage{
empty_plot_for_legend(new = FALSE, ...)
}
\arguments{
\item{new}{Boolean: Put empty plot over existing plotting window?}

\item{...}{any argument that can be passed to \code{par()}, including \code{mfrow}, \code{mar}, and \code{oma}. Only used if \code{new == TRUE}.}
}
\value{
TRUE
}
\description{
Give me an empty plot on top of my \emph{existing} plot so that I can add a legend or other additional annotations.
For example, if I have a multi-panel plot and for some reason can't add subfigure labels to each of them.
The extent of the new plot is c(0,1,0,1)
}
\examples{
empty_plot_for_legend()
}
