% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairy_circle_labels.R
\name{fairy_circle_labels}
\alias{fairy_circle_labels}
\title{Give me map labels in a circle !}
\usage{
fairy_circle_labels(
  sites,
  centre,
  label_radius,
  site_names,
  rank_cutoff = FALSE,
  n_toadstools = 65,
  gap = 0.95,
  guides = FALSE,
  line_col = "black",
  lab_col = "black",
  lab_cex = 1,
  line_lwd = 1
)
}
\arguments{
\item{sites}{A two-column data.frame of longitudes and latitudes of the sites to be labelled, in label order}

\item{centre}{A vector of length 2, giving coordinates of the middle of our fairy circle}

\item{label_radius}{The radius of our circle (e.g., in decimal degrees)}

\item{site_names}{A vector of labels to add, in label order. (e.g., 1:nrow(sites))}

\item{rank_cutoff}{Would you like to label not all of the sites? Either FALSE or a numeric. Defaults to FALSE.}

\item{n_toadstools}{A numeric indicating the number of possible locations for labels on our fairy circle. Should be greater than the number of labels. Defaults to 65.}

\item{gap}{A numeric between 0 and 1 indicating outer endpoint of label line, as proportion of label_radius. Defaults to 0.95.}

\item{guides}{Would you like some guides while you work out where to put your labels? Either TRUE or FALSE. Defaults to FALSE}

\item{line_col}{A valid colour. Defaults to "red".}

\item{lab_col}{A valid colour. Defaults to "red".}
}
\description{
A function to arrange map labels in a circle around a shape. For each label, finds closest point on circle. Should be called after base map is produced.
}
\examples{
fairy_circle_labels(sites[sel,], label_radius = (plot_box[2] - plot_box[1])/2*0.9, centre = c(112.5, 0), site_names = sel, rank_cutoff = FALSE, n_toadstools = 80)
}
\keyword{labels}
