% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdendro.R
\name{ggdendro}
\alias{ggdendro}
\title{Radar / Star / Tree ring plot}
\usage{
ggdendro(data, axes = NULL, id = NULL, axis_range = NULL, scale_axes = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{axes}{vector of strings: names of columns in \code{data} to include as axes}

\item{id}{NULL or string: name of column in \code{data} indicating row IDs}

\item{axis_range}{NULL or vector of 2 numerics: common axis range}

\item{scale_axes}{Boolean. Scale axes to \link{0,1}?}
}
\value{
ggplot object
}
\description{
Unsure if this will be useful but I'd like the option ......
}
\examples{

dat = data.frame(x = rnorm(5), y = rnorm(5), z = rnorm(5), a = rnorm(5)) + 2
ggdendro(axes = c("x", "y", "z", "a"), data = dat)

}
