% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_simple_dist.R
\name{my_simple_dist}
\alias{my_simple_dist}
\title{My simple (Euclidean) dist}
\usage{
my_simple_dist(pos1, pos2)
}
\arguments{
\item{pos1}{vector of length 2}

\item{pos2}{vector of length 2}
}
\value{
numeric distance
}
\description{
My simple (Euclidean) dist
}
\examples{
my_simple_dist(c(0,0), c(1,1))
}
