% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestLand.R
\name{nearestLand}
\alias{nearestLand}
\title{Drag points to nearest non-NA raster cell}
\usage{
nearestLand(points, raster, max_distance)
}
\arguments{
\item{points}{two-column data.frame of coordinates}

\item{raster}{raster.}

\item{max_distance}{numeric. (search radius)}
}
\value{
data.frame
}
\description{
Pinched from seegSDM which hasn't been maintained
https://github.com/SEEG-Oxford/seegSDM/blob/master/R/seegSDM.R
Written for raster - this version converted for terra
(I use this function all the time !)
}
\details{
There is now the search_radius argument in terra::extract, but this doesn't
work if your raster has more than one layer, even if the layers have the same
mask - hence nearestLand()

(Should partition this and other functions into a terra toolkit but leaving
with visualisation toolkit for now)
}
\examples{
nearestLand(dat_df, ras, 10000)
}
