% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosepal.R
\docType{data}
\name{pals}
\alias{pals}
\title{loose palettes}
\format{
\subsection{\code{pals}}{

An object of class \code{palettes_palette} with 1
colour palettes. Use \code{names(pals)} to return all palette names.
}
}
\source{
\url{https://github.com/lu-harr/looseVis}
}
\usage{
pals
}
\description{
Palettes for looseVis
}
\examples{
# Get all palettes by name.
names(pals)

# Plot all palettes.
plot(pals)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\author{
\href{https://github.com/lu-harr}{Lucinda Harrison}
}
\keyword{datasets}
