% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_to_df.R
\name{rast_plot}
\alias{rast_plot}
\title{Wrap a rough shortcut from rast to ggplot}
\usage{
rast_plot(ras)
}
\arguments{
\item{ras}{SpatRaster}
}
\value{
ggplot
}
\description{
I guess I would terra::plot() would have been fine but at least I can add
layers to the object that is returned
}
\examples{
ras <- rast(rnorm(100))
rast_plot(ras)
}
