% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_to_df.R
\name{rast_to_df}
\alias{rast_to_df}
\title{Prepare rast (stack) for ggplot}
\usage{
rast_to_df(ras)
}
\arguments{
\item{ras}{SpatRaster}
}
\value{
data.frame with columns "x", "y", "lyr", "val"
}
\description{
Prepare rast (stack) for ggplot
}
\examples{
ras <- rast(rnorm(100))
df <- rast_to_df(ras)
}
