% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cex.R
\name{scale_cex}
\alias{scale_cex}
\title{Scale value for point size (for example)}
\usage{
scale_cex(from, trans = NULL, min_cex = 0.5, max_cex = 1)
}
\arguments{
\item{from}{vector of numerics}

\item{trans}{NULL or function that takes a numeric as input, e.g., \code{log10}, \code{exp}, \code{sqrt}, ...}

\item{min_cex}{numeric, >= 0 & <= \code{max_cex}}

\item{max_cex}{numeric, >= 0 & >= \code{min_cex}}
}
\value{
vector of transformed data
}
\description{
A function to transform a vector of numerics and return the data scaled
between \code{min_cex} and \code{max_cex}. Useful for scaling point sizes and, I assume,
other things.
}
\examples{
x <- rexp(100)
scaled_cex <- scale_cex(x, log10)
}
