% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfigure_label.R
\name{subfigure_label}
\alias{subfigure_label}
\title{Give me a subfigure label}
\usage{
subfigure_label(
  plot_region,
  x_displacement,
  y_displacement,
  label,
  cex.label = 1
)
}
\arguments{
\item{plot_region}{coordinates of plotting region, e.g. par()$usr}

\item{x_displacement}{numeric between 0 and 1 indicating proportional x displacement (0 for far left, 1 for far right)}

\item{y_displacement}{numeric between 0 and 1 indicating proportional y displacement (0 for bottom, 1 for top)}

\item{label}{character vector to add to plot}

\item{cex.label}{numeric to control label size. Defaults to 1.}
}
\value{
TRUE
}
\description{
Give me a subfigure label
}
\examples{
subfigure_label(par()$usr, 0.1, 0.9, "(a)", 1.3)
}
